购物车

+++ app/src/main/java/maoyan/pjs/com/maoyan/bean/WaitListBean.java
@@ -114,7 +114,7 @@ public class WaitListBean {
             private String img;
             private boolean late;
             private int localPubSt;
-            private int mk;
+            private double mk;
             private String nm;
             private int pn;
             private boolean preShow;
@@ -123,7 +123,7 @@ public class WaitListBean {
             private int recentShowDate;
             private int recentShowNum;
             private String rt;
-            private int sc;
+            private double sc;
             private String scm;
             private String showInfo;
             private int showNum;
@@ -234,11 +234,11 @@ public class WaitListBean {
                 this.localPubSt = localPubSt;
             }
 
-            public int getMk() {
+            public double getMk() {
                 return mk;
             }
 
-            public void setMk(int mk) {
+            public void setMk(double mk) {
                 this.mk = mk;
             }
 
@@ -306,11 +306,11 @@ public class WaitListBean {
                 this.rt = rt;
             }
 
-            public int getSc() {
+            public double getSc() {
                 return sc;
             }
 
-            public void setSc(int sc) {
+            public void setSc(double sc) {
                 this.sc = sc;
             }
